updated testing

+++ Backend/src/main/java/idealab/api/schedule/DeleteUserInfoCronJob.java
@@ -40,32 +40,30 @@ public class DeleteUserInfoCronJob {
         Iterable<CustomerInfo> customerInfoList = customerInfoRepo.findAll();
         List<CustomerInfoCronJobInfo> customersToDelete = new ArrayList<>();
 
+        if (customerInfoList == null)
+            return;
+
         customerInfoList.forEach(c -> {
             customersToDelete.add(new CustomerInfoCronJobInfo(c));
         });
 
         LocalDateTime now = LocalDateTime.now();
-        if (customersToDelete != null)
-        {
-            for (CustomerInfoCronJobInfo c : customersToDelete)
-            {
-                if(c.getCreatedAt().plus(NUM_DAYS_RETENTION, ChronoUnit.DAYS).isBefore(now))
-                {
-                    List<PrintJob> pJobs = printJobRepo.findNewestPrintJobByCustomerId(c.getId());
-                    PrintJob p = pJobs.get(0);
-                    if(p != null)
-                    {
-                        if(p.getCreatedAt().plus(NUM_DAYS_RETENTION, ChronoUnit.DAYS).isBefore(now))
-                        {
-                            printJobRepo.deleteAll(pJobs);
-                            customerInfoRepo.deleteById(c.getId());
+        if (customersToDelete != null) {
+            for (CustomerInfoCronJobInfo c : customersToDelete) {
+                if (c.getCreatedAt().plus(NUM_DAYS_RETENTION, ChronoUnit.DAYS).isBefore(now)) {
+                    List<PrintJob> pJobs = printJobRepo.findPrintJobsByCustomerIdNewestFirst(c.getId());
+                    if (pJobs != null && pJobs.size() >= 1) {
+                        PrintJob p = pJobs.get(0);
+                        if (p != null) {
+                            if (p.getCreatedAt().plus(NUM_DAYS_RETENTION, ChronoUnit.DAYS).isBefore(now)) {
+                                printJobRepo.deleteAll(pJobs);
+                                customerInfoRepo.deleteById(c.getId());
+                            }
                         }
                     }
                 }
             }
-        }
-        else
-        {
+        } else {
             LOGGER.info("No users to delete");
         }
     }
